+++ src/frequenz/sdk/timeseries/_base_types.py
@@ -4,7 +4,7 @@
 """Timeseries basic types."""
 
 import functools
-from dataclasses import dataclass, field
+from dataclasses import dataclass
 from datetime import datetime, timezone
 from typing import Callable, Iterator, Self, overload
 
@@ -12,10 +12,6 @@ UNIX_EPOCH = datetime.fromtimestamp(0.0, tz=timezone.utc)
 """The UNIX epoch (in UTC)."""
 
 
-# Ordering by timestamp is a bit arbitrary, and it is not always what might be
-# wanted. We are using this order now because usually we need to do binary
-# searches on sequences of samples, and the Python `bisect` module doesn't
-# support providing a key until Python 3.10.
 @dataclass(frozen=True, order=True)
 class Sample:
     """A measurement taken at a particular point in time.
@@ -25,10 +21,10 @@ class Sample:
     coherent view on a group of component metrics for a particular timestamp.
     """
 
-    timestamp: datetime = field(compare=True)
+    timestamp: datetime
     """The time when this sample was generated."""
 
-    value: float | None = field(compare=False, default=None)
+    value: float | None = None
     """The value of this sample."""